# docs/api/v2/V2_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API explorer for the go-deploy API. You can use it as a reference for the API endpoints.",
        "title": "go-deploy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "https://github.com/kthcloud/go-deploy"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/kthcloud/go-deploy?tab=MIT-1-ov-file#readme"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v2/resourceMigrations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List resource migrations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ResourceMigration"
                ],
                "summary": "List resource migrations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/body.ResourceMigrationRead"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create resource migration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ResourceMigration"
                ],
                "summary": "Create resource migration",
                "parameters": [
                    {
                        "description": "Resource Migration Create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body.ResourceMigrationCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.ResourceMigrationCreated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/resourceMigrations/{resourceMigrationId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get resource migration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ResourceMigration"
                ],
                "summary": "Get resource migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource Migration ID",
                        "name": "resourceMigrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.ResourceMigrationRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update resource migration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ResourceMigration"
                ],
                "summary": "Update resource migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource Migration ID",
                        "name": "resourceMigrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resource Migration Update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body.ResourceMigrationUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.ResourceMigrationUpdated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete resource migration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ResourceMigration"
                ],
                "summary": "Delete resource migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource Migration ID",
                        "name": "resourceMigrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "body.ResourceMigrationCreate": {
            "type": "object",
            "required": [
                "resourceId",
                "type"
            ],
            "properties": {
                "resourceId": {
                    "description": "ResourceID is the ID of the resource that is being migrated.\nThis can be a VM ID, deployment ID, etc. depending on the type of the migration.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the resource migration.\nIt is used by privileged admins to directly accept or reject a migration.\nThe field is ignored by non-admins.\n\nPossible values:\n- accepted\n- pending",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the type of the resource migration.\n\nPossible values:\n- updateOwner",
                    "type": "string",
                    "enum": [
                        "updateOwner"
                    ]
                },
                "updateOwner": {
                    "description": "UpdateOwner is the set of parameters that are required for the updateOwner migration type.\nIt is ignored if the migration type is not updateOwner.",
                    "type": "object",
                    "required": [
                        "ownerId"
                    ],
                    "properties": {
                        "ownerId": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "body.ResourceMigrationCreated": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "description": "JobID is the ID of the job that was created for the resource migration.\nIt will only be set if the migration was created with status 'accepted'.",
                    "type": "string"
                },
                "resourceId": {
                    "description": "ResourceID is the ID of the resource that is being migrated.\nThis can be a VM ID, deployment ID, etc. depending on the type of the migration.",
                    "type": "string"
                },
                "resourceType": {
                    "description": "ResourceType is the type of the resource that is being migrated.\n\nPossible values:\n- vm\n- deployment",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the resource migration.\nWhen this field is set to 'accepted', the migration will take place and then automatically be deleted.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the type of the resource migration.\n\nPossible values:\n- updateOwner",
                    "type": "string"
                },
                "updateOwner": {
                    "description": "UpdateOwner is the set of parameters that are required for the updateOwner migration type.\nIt is empty if the migration type is not updateOwner.",
                    "type": "object",
                    "properties": {
                        "ownerId": {
                            "type": "string"
                        }
                    }
                },
                "userId": {
                    "description": "UserID is the ID of the user who initiated the migration.",
                    "type": "string"
                }
            }
        },
        "body.ResourceMigrationRead": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resourceId": {
                    "description": "ResourceID is the ID of the resource that is being migrated.\nThis can be a VM ID, deployment ID, etc. depending on the type of the migration.",
                    "type": "string"
                },
                "resourceType": {
                    "description": "ResourceType is the type of the resource that is being migrated.\n\nPossible values:\n- vm\n- deployment",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the resource migration.\nWhen this field is set to 'accepted', the migration will take place and then automatically be deleted.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the type of the resource migration.\n\nPossible values:\n- updateOwner",
                    "type": "string"
                },
                "updateOwner": {
                    "description": "UpdateOwner is the set of parameters that are required for the updateOwner migration type.\nIt is empty if the migration type is not updateOwner.",
                    "type": "object",
                    "properties": {
                        "ownerId": {
                            "type": "string"
                        }
                    }
                },
                "userId": {
                    "description": "UserID is the ID of the user who initiated the migration.",
                    "type": "string"
                }
            }
        },
        "body.ResourceMigrationUpdate": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "code": {
                    "description": "Code is a token required when accepting a migration if the acceptor is not an admin.\nIt is sent to the acceptor using the notification API",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the resource migration.\nIt is used to accept a migration by setting the status to 'accepted'.\nIf the acceptor is not an admin, a Code must be provided.\n\nPossible values:\n- accepted\n- pending",
                    "type": "string"
                }
            }
        },
        "body.ResourceMigrationUpdated": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "description": "JobID is the ID of the job that was created for the resource migration.\nIt will only be set if the migration was updated with status 'accepted'.",
                    "type": "string"
                },
                "resourceId": {
                    "description": "ResourceID is the ID of the resource that is being migrated.\nThis can be a VM ID, deployment ID, etc. depending on the type of the migration.",
                    "type": "string"
                },
                "resourceType": {
                    "description": "ResourceType is the type of the resource that is being migrated.\n\nPossible values:\n- vm\n- deployment",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the resource migration.\nWhen this field is set to 'accepted', the migration will take place and then automatically be deleted.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the type of the resource migration.\n\nPossible values:\n- updateOwner",
                    "type": "string"
                },
                "updateOwner": {
                    "description": "UpdateOwner is the set of parameters that are required for the updateOwner migration type.\nIt is empty if the migration type is not updateOwner.",
                    "type": "object",
                    "properties": {
                        "ownerId": {
                            "type": "string"
                        }
                    }
                },
                "userId": {
                    "description": "UserID is the ID of the user who initiated the migration.",
                    "type": "string"
                }
            }
        },
...
        "sys.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "sys.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sys.Error"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        }
    }
}